+++ test/full_coverage_test.dart
@@ -79,6 +79,8 @@ import 'package:pi_hole_client/repository/database.dart';
 import 'package:pi_hole_client/repository/secure_storage.dart';
 import 'package:pi_hole_client/screens/app_logs/app_log_details_modal.dart';
 import 'package:pi_hole_client/screens/app_logs/app_logs.dart';
+import 'package:pi_hole_client/screens/common/empty_data_screen.dart';
+import 'package:pi_hole_client/screens/common/error_message_screen.dart';
 import 'package:pi_hole_client/screens/domains/add_domain_modal.dart';
 import 'package:pi_hole_client/screens/domains/delete_domain_modal.dart';
 import 'package:pi_hole_client/screens/domains/domain_comment_modal.dart';
@@ -125,6 +127,7 @@ import 'package:pi_hole_client/screens/settings/app_settings/language_screen.dar
 import 'package:pi_hole_client/screens/settings/app_settings/theme_screen.dart';
 import 'package:pi_hole_client/screens/settings/server_settings/server_info.dart';
 import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/host_information_section.dart';
+import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/info_trailing.dart';
 import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/performance_usage_section.dart';
 import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/pihole_version_section.dart';
 import 'package:pi_hole_client/screens/settings/server_settings/widgets/server_info/server_connection_section.dart';
@@ -144,8 +147,6 @@ import 'package:pi_hole_client/widgets/bottom_nav_bar.dart';
 import 'package:pi_hole_client/widgets/custom_list_tile.dart';
 import 'package:pi_hole_client/widgets/custom_radio.dart';
 import 'package:pi_hole_client/widgets/custom_radio_list_tile.dart';
-import 'package:pi_hole_client/widgets/empty_data_screen.dart';
-import 'package:pi_hole_client/widgets/error_message_screen.dart';
 import 'package:pi_hole_client/widgets/list_tile_title.dart';
 import 'package:pi_hole_client/widgets/navigation_rail.dart';
 import 'package:pi_hole_client/widgets/numeric_pad.dart';
